pipeline only runs on valid files in raw dir
